Move discovery logic to its own file
